Update 1.00.4
Added Chord progression Functionality and switched from Update() to FixedUpdate(). I Switched from time based results to an incrementing integer. I made it so that if the seed is equal to 0 then it will randomly generate a seed. Since we will be changing the key that we are in frequently, I didnt want any big jumps trying to transpose. Thus I reordered the scales in Int2Note.cs so that they start off with either A or A#. I also made it so the start of a new number category will be C not A. The waitforseconds was buggy so I set it to a constant and now its working for some reason. The audio player is still kinda mid so i recommend just importing the notes into basca ceoil or some other music software. In an attempt to create christmas music I accidentally created Jazz so thats in the examples folder if you want to listen to it.